* Add .vscode include paths

* Update whitewhale firmware submodule to newer libavr32

* Update for breaking changes in the Rack 0.6 API

* Restore arch.mk include in Makefile

* Stop filtering -MMD flag out in Makefile

* Package firmware in res folder for dist build, remove -o0 flag, correct some build warnings

* Update 0.6 build for OSX

+++ src/common/MonomeModuleBase.cpp
@@ -40,7 +40,7 @@ void MonomeModuleBase::setGridConnection(GridConnection* newConnection)
 
         string id = gridConnection->device->id;
         string wide_id;
-        for (int i = 0; i < id.length(); i++)
+        for (size_t i = 0; i < id.length(); i++)
         {
             if (i == 1 && id[1] == 'v')
                 i++;
